requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
conda==4.3.16
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.2
django-allauth==0.43.0
django-smart-selects==1.5.9
filelock==3.0.12
idna==2.10
oauthlib==3.1.0
pipenv==2020.8.13
pycosat==0.6.3
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
